+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1] - 2021-08-26
 ### Fixed
 - Support `?` inside bound arguments, [#18](https://github.com/loyd/clickhouse.rs/issues/18).
-- Use the `POST` method if a query is bigger than 8KiB [#19](https://github.com/loyd/clickhouse.rs/issues/19).
+- Use the `POST` method if a query is bigger than 8KiB, [#19](https://github.com/loyd/clickhouse.rs/issues/19).
 
 ## [0.8.0] - 2021-07-28
 ### Fixed
@@ -118,7 +119,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client::insert()` for inserting into tables.
 - `Client::query()` for selecting from tables and DDL statements.
 
-[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.8.1...HEAD
+[0.8.1]: https://github.com/loyd/clickhouse.rs/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/loyd/clickhouse.rs/compare/v0.7.2...v0.8.0
 [0.7.2]: https://github.com/loyd/clickhouse.rs/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/loyd/clickhouse.rs/compare/v0.7.0...v0.7.1